+++ src/integrationTest/java/uk/gov/hmcts/reform/civil/controllers/dashboard/scenarios/ClaimFeeRequiredScenarioTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.civil.controllers.dashboard.scenarios.claimant;
+package uk.gov.hmcts.reform.civil.controllers.dashboard.scenarios;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
@@ -15,9 +15,9 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @Testcontainers
-public class HwfRequestedScenarioTest extends BaseIntegrationTest {
+public class ClaimFeeRequiredScenarioTest extends BaseIntegrationTest {
 
-    public static final String SCENARIO_HWF_REQUESTED = "Scenario.AAA7.ClaimIssue.HWF.Requested";
+    public static final String SCENARIO_CLAIM_FEE_REQUIRED = "Scenario.AAA7.ClaimIssue.ClaimFee.Required";
     private static final String DASHBOARD_CREATE_SCENARIO_URL
         = "/dashboard/scenarios/{scenario_ref}/{unique_case_identifier}";
     private static final String GET_NOTIFICATIONS_URL
@@ -25,13 +25,14 @@ public class HwfRequestedScenarioTest extends BaseIntegrationTest {
     private static final String GET_TASKS_ITEMS_URL = "/dashboard/taskList/{ccd-case-identifier}/role/{role-type}";
 
     @Test
-    void should_create_scenario_hwf_requested() throws Exception {
+    void should_create_scenario_for_claim_fee_required() throws Exception {
 
         UUID caseId = UUID.randomUUID();
         doPost(BEARER_TOKEN,
                ScenarioRequestParams.builder()
-                   .params(Map.of()).build(),
-               DASHBOARD_CREATE_SCENARIO_URL, SCENARIO_HWF_REQUESTED, caseId
+                   .params(Map.of("claimFee", "£70"
+                   )).build(),
+               DASHBOARD_CREATE_SCENARIO_URL, SCENARIO_CLAIM_FEE_REQUIRED, caseId
         )
             .andExpect(status().isOk());
 
@@ -46,8 +47,8 @@ public class HwfRequestedScenarioTest extends BaseIntegrationTest {
                 jsonPath("$[1].currentStatusEn").value("Available"),
                 jsonPath("$[2].taskNameEn").value("<a>View the response to the claim</a>"),
                 jsonPath("$[2].currentStatusEn").value("Not available yet"),
-                jsonPath("$[3].taskNameEn").value("<a href={VIEW_INFO_ABOUT_DEFENDANT_URL}  rel=\"noopener noreferrer\" class=\"govuk-link\">View information about the defendant</a>"),
-                jsonPath("$[3].currentStatusEn").value("Available"),
+                jsonPath("$[3].taskNameEn").value("<a>View information about the defendant</a>"),
+                jsonPath("$[3].currentStatusEn").value("Not available yet"),
                 jsonPath("$[4].taskNameEn").value("<a>View hearings</a>"),
                 jsonPath("$[4].currentStatusEn").value("Not available yet"),
                 jsonPath("$[5].taskNameEn").value("<a>Upload hearing documents</a>"),
@@ -72,8 +73,8 @@ public class HwfRequestedScenarioTest extends BaseIntegrationTest {
             .andExpect(status().isOk())
             .andExpectAll(
                 status().is(HttpStatus.OK.value()),
-                jsonPath("$[0].titleEn").value("We’re reviewing your help with fees application"),
-                jsonPath("$[0].descriptionEn").value("You’ve applied for help with the claim fee. You’ll receive an update in 5 to 10 working days.")
+                jsonPath("$[0].titleEn").value("You need to pay your claim fee"),
+                jsonPath("$[0].descriptionEn").value("Your claim has not yet been issued, in order to proceed you must pay the claim fee of £70. <a href={CLAIM_FEE_URL}  rel=\"noopener noreferrer\" class=\"govuk-link\">Pay the claim fee</a>.")
             );
     }
