+++ src/main/java/uk/gov/hmcts/reform/civil/callback/CaseEvent.java
@@ -364,8 +364,7 @@ public enum CaseEvent {
     NOTIFY_LIP_CLAIMANT_HWF_OUTCOME(CAMUNDA),
     GENERATE_DASHBOARD_NOTIFICATION_CUI(CAMUNDA),
     CREATE_DASHBOARD_NOTIFICATION_FOR_CLAIM_ISSUE_FOR_APPLICANT1(CAMUNDA),
-    CREATE_DASHBOARD_NOTIFICATION_FOR_CLAIM_ISSUE_FOR_RESPONDENT1(CAMUNDA),
-    GENERATE_DASHBOARD_NOTIFICATION_CLAIM_ISSUE_HWF_CLAIMANT1(CAMUNDA);
+    CREATE_DASHBOARD_NOTIFICATION_FOR_CLAIM_ISSUE_FOR_RESPONDENT1(CAMUNDA);
 
     private final UserType userType;
