+++ src/main/resources/db/migration/V2024_02_02_12356_1__create_dashboard_notification_template_table.sql
@@ -7,8 +7,8 @@ CREATE TABLE IF NOT EXISTS dbs.dashboard_notifications_templates (
  template_name VARCHAR(256),
  title_en VARCHAR(256),
  title_cy VARCHAR(256),
- description_en VARCHAR(512),
- description_cy VARCHAR(512),
+ description_en VARCHAR(256),
+ description_cy VARCHAR(256),
  notification_role VARCHAR(256),
  time_to_live VARCHAR(256),
  created_at TIMESTAMP default CURRENT_TIMESTAMP
