--- src/main/resources/db/migration/V2024_02_29_1420__CIV-12784_claim_issue_invalid_hwf_scenario.sql
@@ -1,18 +1,0 @@
-/**
- * Add scenario
- */
-INSERT INTO dbs.scenario (name, notifications_to_delete, notifications_to_create)
-VALUES ('Scenario.AAA7.ClaimIssue.HWF.InvalidRef', '{"Notice.AAA7.ClaimIssue.HWF.Requested"}',
-        '{"Notice.AAA7.ClaimIssue.HWF.InvalidRef" : ["typeOfFee"]}');
-
-/**
- * Add notification template
- */
-INSERT INTO dbs.dashboard_notifications_templates (template_name, title_En, title_Cy, description_En, description_Cy
-                                                  ,notification_role)
-VALUES ('Notice.AAA7.ClaimIssue.HWF.InvalidRef',
-        'You''ve provided an invalid help with fees reference number',
-        'You''ve provided an invalid help with fees reference number',
-        'You''ve applied for help with the ${typeOfFee} fee, but the reference number is invalid.<br>You''ve been sent an email with instructions on what to do next. If you''ve already read the email and taken action, disregard this message.<br>You can pay by phone by calling {civilMoneyClaimsTelephone}.',
-        'You''ve applied for help with the ${typeOfFee} fee, but the reference number is invalid.<br>You''ve been sent an email with instructions on what to do next. If you''ve already read the email and taken action, disregard this message.<br>You can pay by phone by calling {civilMoneyClaimsTelephone}.',
-           'CLAIMANT');
